> Fixed minor unit conversion error in PlayerController.cs. > Fixed parameter issue in BuildSetUp.unity. > Fixed time scale issue with code in WarmUpPortalController.cs. > Added Rewired prefab to prefabs.

+++ TrialsOfTheRiftWC/Assets/Scripts/PlayerController.cs
@@ -121,7 +121,7 @@ public class PlayerController : MonoBehaviour{
         if (!isWisp)
         {
             print("heal" + heal);
-            int tempHp = f_playerHealth + heal;
+            int tempHp = (int)(f_playerHealth + heal);
             if (tempHp >= Constants.PlayerStats.C_MaxHealth)
             {
                 f_playerHealth = Constants.PlayerStats.C_MaxHealth;
